Move Weaviate sync logic to RagLogic
